move range overview layout to a separate file
