+++ console/src/range/external.ts
@@ -9,10 +9,12 @@
 
 import { type Layout } from "@/layout";
 import { Create, CREATE_LAYOUT_TYPE } from "@/range/CreateLayout";
-import { Overview, OVERVIEW_LAYOUT } from "@/range/overview/Overview";
+import { OVERVIEW_LAYOUT } from "@/range/overview/layout";
+import { Overview } from "@/range/overview/Overview";
 
 export * from "@/range/ContextMenu";
 export * from "@/range/CreateLayout";
+export * from "@/range/overview/layout";
 export * from "@/range/overview/Overview";
 export * from "@/range/Select";
 export * from "@/range/selectors";
